Diego

+++ SMedia/SMedia/Models/Reaction.cs
@@ -7,10 +7,10 @@ namespace SMedia.Models
 {
     public class Reaction
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public byte Type { get; set; }
-        public int UserId { get; set; }
-        public int PostId { get; set; }
+        public long UserId { get; set; }
+        public long PostId { get; set; }
 
         public virtual User User { get; set; }
         public virtual Post Post { get; set; }
